added handling for math blocks in rule heads
The result of SPARQL math functions can now be assigned to a variable
in a rule head using the math-blocks implemented here. An example of
the syntax is: [:float :as ?/a_over_b ?/a "/" ?/b]. In this case the
resultant SPARQL is "(xsd:float(?a)/xsd:float(?b) AS ?a_over_b)"  The
first argument of the vector defines the numeric type that will be
assigned to all variables. It can be :float, :double, :integer, or
:decimal.  The :as serves no purpose other than perhaps to serve as a
key to the reader that the next variable (?/a_over_b in this case) will
be used to assign the result. All arguments after the result variable
are treated as part of the equation. See the unit tests for some more
examples.