reformatted math block syntax to use a map instead of a vector
While adding handling for SPARQL "group by" clauses it became apparent
that the math block would be more easy to work with if it were a map.
Math blocks are now maps with keys:  :num_type to denote the number
type to use, e.g. :float; :as to denote the result variable; :eqn to
denote the equation; and optionally :group_by to denote variables that
should be listed in a SPARQL "group by" clause. See unit tests for
details and examples. In the future, the math-block can serve as a
general "function-block" and can absorb the count-block functionality.
Doing so would make the code and rule syntax simpler.