Revised the classpath-matching function to allow for filtering by file suffix; Rule filtering now mandates .clj suffix
This function is used to extract rule files from the classpath. It
previously included any file in the classpath that matches the input
path. In order to prevent inclusion of rule documentation (e.g. .png
and other files), this function now allows a file suffix to be
specified and used as part of the classpath match.