Revised register-namespaces
Split the reduce into two separate cases so that the try/catch block
for UnsupportedOperationException could be moved to this level.
Previously there was an error message displayed each time a namespace
was attempted to be registered with the kb. Now there should be just a
single error message indicating the UnsupportedOperationException.